pipeline/requirements.txt
rembg
Pillow
numpy
boto3
requests
supabase
torch
torchvision
ffmpeg-python
